+++ src/components/progress-tracker/progress-tracker.d.ts
@@ -1,6 +1,19 @@
 import { MarginProps } from "styled-system";
 
 export interface ProgressBarProps extends MarginProps {
+  /** Specifies an aria-label to the component */
+  "aria-label"?: string;
+  /** Specifies the aria-describedby for the component */
+  "aria-describedby"?: string;
+  /** The value of progress to be read out to the user. */
+  "aria-valuenow"?: number;
+  /** The minimum value of the progress tracker */
+  "aria-valuemin"?: number;
+  /** The maximum value of the progress tracker */
+  "aria-valueMax"?: number;
+  /** Prop to define the human readable text alternative of aria-valuenow
+   * if aria-valuenow is not a number */
+  "aria-valuetext"?: string;
   /** Size of the progressBar. */
   size?: "small" | "medium" | "large";
   /** Current progress (percentage). */